Move old string built-in methods to internal.
