+++ stdlib/internal/src/main/java/org/ballerinalang/stdlib/internal/builtin/ToByteArray.java
@@ -15,14 +15,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.langlib.string;
+package org.ballerinalang.stdlib.internal.builtin;
 
-import org.ballerinalang.bre.Context;
-import org.ballerinalang.bre.bvm.BlockingNativeCallableUnit;
 import org.ballerinalang.jvm.Strand;
 import org.ballerinalang.jvm.values.ArrayValue;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BValueArray;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.ReturnType;
@@ -35,27 +32,14 @@ import java.io.UnsupportedEncodingException;
  * Convert String to byte array.
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "lang.string",
+        orgName = "ballerina", packageName = "internal",
         functionName = "toByteArray",
         args = {@Argument(name = "string", type = TypeKind.STRING),
                 @Argument(name = "encoding", type = TypeKind.STRING)},
         returnType = {@ReturnType(type = TypeKind.ARRAY, elementType = TypeKind.BYTE)},
         isPublic = true
 )
-public class ToByteArray extends BlockingNativeCallableUnit {
-
-    public void execute(Context ctx) {
-        try {
-            String string = ctx.getStringArgument(0);
-            String encoding = ctx.getStringArgument(1);
-            byte[] bytes = string.getBytes(encoding);
-            BValueArray byteArray = new BValueArray(bytes);
-            ctx.setReturnValues(byteArray);
-        } catch (UnsupportedEncodingException e) {
-            throw new BallerinaException(BallerinaErrorReasons.STRING_OPERATION_ERROR,
-                                         "Unsupported Encoding " + e.getMessage());
-        }
-    }
+public class ToByteArray {
 
     public static ArrayValue toByteArray(Strand strand, String value, String encoding) {
         StringUtils.checkForNull(value, encoding);
