+++ stdlib/internal/src/main/java/org/ballerinalang/stdlib/internal/builtin/Unescape.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.ballerinalang.stdlib.internal.builtin;
+
+import org.ballerinalang.jvm.Strand;
+import org.ballerinalang.model.types.TypeKind;
+import org.ballerinalang.natives.annotations.Argument;
+import org.ballerinalang.natives.annotations.BallerinaFunction;
+import org.ballerinalang.natives.annotations.ReturnType;
+
+/**
+ * Extern function ballerina.model.strings:unescape.
+ *
+ * @since 0.8.0
+ */
+@BallerinaFunction(
+        orgName = "ballerina", packageName = "internal",
+        functionName = "unescape",
+        args = {@Argument(name = "s", type = TypeKind.STRING)},
+        returnType = {@ReturnType(type = TypeKind.STRING)},
+        isPublic = true
+)
+public class Unescape {
+
+    public static String unescape(Strand strand, String value) {
+        StringUtils.checkForNull(value);
+        return value.replace("\\", "");
+    }
+}
